minor  update

+++ examples/[template][bare-metal]/project/mdk/RTE/Acceleration/arm_2d.c
@@ -19,10 +19,10 @@
 /* ----------------------------------------------------------------------
  * Project:      Arm-2D Library
  * Title:        arm-2d.c
- * Description:  Tables for pixel pipeline OPs
+ * Description:  Essential components of Arm-2D
  *
- * $Date:        19. Nov 2023
- * $Revision:    V.1.3.1
+ * $Date:        4. April 2024
+ * $Revision:    V.1.3.2
  *
  * Target Processor:  Cortex-M cores
  *
@@ -236,7 +236,7 @@ void __arm_2d_log_printf(int32_t nIndentLevel,
         s_wLineNumber++;
 
         /* start a new line */
-        __ARM_2D_PORT_PRINTF__("\r\n[%010d]\t", (int)s_wLineNumber);
+        __ARM_2D_PORT_PRINTF__("\r\n[%010d]\t", s_wLineNumber);
 
         for (int32_t n = 0; n < nIndentLevel; n++) {
             __ARM_2D_PORT_PRINTF__("\t");
